♻️ Case entity -> modern
