+++ src/packages/@ncigdc/modern_components/AddOrRemoveAllFilesButton/index.js
@@ -1,0 +1,11 @@
+import PropTypes from 'prop-types';
+import Component from './AddOrRemoveAllFilesButton';
+import createRendererBtn from './AddOrRemoveAllFilesButton.relay';
+import createRendererFilesTotal from './CaseFilesTotal.relay';
+const AddOrRemoveAllFilesButton = createRendererFilesTotal(
+  createRendererBtn(Component),
+);
+AddOrRemoveAllFilesButton.propTypes = {
+  caseId: PropTypes.string,
+};
+export default AddOrRemoveAllFilesButton;
