+++ src/packages/@ncigdc/modern_components/BiospecimenCard/index.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+import Component from './BiospecimenCard';
+import createRenderer from './BiospecimenCard.relay';
+const BiospecimenCard = createRenderer(Component);
+BiospecimenCard.propTypes = {
+  caseId: PropTypes.string,
+};
+export default BiospecimenCard;
