+++ src/packages/@ncigdc/modern_components/CaseCounts/index.js
@@ -1,0 +1,8 @@
+import CaseCountsDataCategoryComponent from './CaseCountsDataCategory';
+import CaseCountsExpStrategyComponent from './CaseCountsExpStrategy';
+import createRenderer from './CaseCounts.relay';
+
+const CaseCountsDataCategory = createRenderer(CaseCountsDataCategoryComponent);
+const CaseCountsExpStrategy = createRenderer(CaseCountsExpStrategyComponent);
+
+export { CaseCountsDataCategory, CaseCountsExpStrategy };
