+++ src/packages/@ncigdc/modern_components/CaseSummary/index.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+import Component from './CaseSummary';
+import createRenderer from './CaseSummary.relay';
+const CaseSummary = createRenderer(Component);
+CaseSummary.propTypes = {
+  caseId: PropTypes.string,
+};
+export default CaseSummary;
