+++ src/packages/@ncigdc/utils/trace.js
@@ -27,10 +27,8 @@ export default (additionalMessage = '') => Component =>
           : ''}%ctracing ${Component.displayName} props:\n${'tracing '.replace(
           /./g,
           '=',
-        )}${Component.displayName.replace(/./g, '=')}=${'props:'.replace(
-          /./g,
-          '=',
-        )}`,
+        )}${(Component.displayName || 'unnamed component')
+          .replace(/./g, '=')}=${'props:'.replace(/./g, '=')}`,
         'color: rgb(22, 138, 96);font-weight: bold;',
         props,
       ),
